+++ src/Enums/CacheKey.php
@@ -8,7 +8,4 @@ enum CacheKey: string
 {
     case PERMISSIONS = 'guard_permissions';
     case ROLES = 'guard_roles';
-    case USER_PERMISSIONS = 'guard_user_permissions_';
-    case USER_ROLES = 'guard_user_roles_';
-    case ROLE_PERMISSIONS = 'guard_role_permissions_';
 }